Move first-load-tracker into performance-tracking module
