+++ client/sites/v2/router.tsx
@@ -1,7 +1,7 @@
 import { siteBySlugQuery, queryClient } from '@automattic/api-queries';
 import page from '@automattic/calypso-router';
 import { Outlet, createRootRouteWithContext, createRoute } from '@tanstack/react-router';
-import { consumeFirstLoad } from 'calypso/dashboard/app/router/first-load-tracker';
+import { peekFirstLoad } from 'calypso/dashboard/app/performance-tracking';
 import { canManageSite } from 'calypso/dashboard/sites/features';
 import { getSiteDisplayName } from 'calypso/dashboard/utils/site-name';
 import { hasSiteTrialEnded } from 'calypso/dashboard/utils/site-trial';
@@ -15,7 +15,7 @@ import type { RootRouterContext } from 'calypso/dashboard/app/router/root';
 export const rootRoute = createRootRouteWithContext< RootRouterContext >()( {
 	component: Root,
 	beforeLoad: async ( { cause } ): Promise< { fullPageLoad: boolean } > => ( {
-		fullPageLoad: cause === 'enter' && consumeFirstLoad(),
+		fullPageLoad: cause === 'enter' && peekFirstLoad(),
 	} ),
 } );
